Create Message<T> class
This class is intended to represent the messages exchanged between servers.
It contains the port to identify the server adrress. It contains the object
that the application wants to exchange and it contains a vector clock representing
the sender vector clock when he sent the message. This clock is important for
the causal delivery process.