Add causal delivery tests
When receiving a message from a server it's important to check if
it passes on the causal tests, otherwise it has to be cached for a
later analysis.

Change the access policy for AsynchronousProcess, CausalDelivery and
Message, because in the future this classes will be inside a package
where the only class accessed shall be the Facade.