Causal delivery (#12)
* Remove Connectivity Test

Remove Connectivity Test from AsynchronousProcess because it was only
created to test the first code (Clique Network) and there is no need to
continue having that code since it's not important anymore.

* Create Message<T> class

This class is intended to represent the messages exchanged between servers.
It contains the port to identify the server adrress. It contains the object
that the application wants to exchange and it contains a vector clock representing
the sender vector clock when he sent the message. This clock is important for
the causal delivery process.

* Add causal delivery tests

When receiving a message from a server it's important to check if
it passes on the causal tests, otherwise it has to be cached for a
later analysis.

Change the access policy for AsynchronousProcess, CausalDelivery and
Message, because in the future this classes will be inside a package
where the only class accessed shall be the Facade.

* Interconnect AsynchronousProcess and CausalDelivery classes

Change just a few initialization parameters, in order to interconnect
both classes.

* Create Middleware package

This package offers an encapsulated and interchangeable layer, that is
responsable for send an receive messages attending to causal delivery.

* Add a Facade to Middleware package

This Facade will be accessed from the app and from the CausalDelivery class.

* Add routine for checking pending messages

This routine is executed everytime that a message arrives because it's important
to see if the arriving message can unlock any message in the queue.

* Add the sending messages logic

When sending a message the event counter has to increment so the clock can proceed.

Add a little chat Test in the MainTester to prove the good working of the middleware
layer.